🍂🗼 Checkpoint ./cmd/addmoney.js:8710627/1220 ./cmd/configvip.js:8710627/4863 ./cmd/delimg.js:8710627/31

+++ cmd/delimg.js
@@ -11,6 +11,8 @@ const vips_db = new db.crearDB("Vips");
 .setColor("RED")
 )
     const img_db = new db.crearDB("img");
+  
+  if(!)
 
     img_db.delete(`${message.author.id}`, args.join(" "));
